clearing

+++ src/engine/input.c
@@ -6,7 +6,7 @@ static	char	*add_to_str(char *str, char c)
 	char	*res;
 
 	res = (char*)malloc((ft_strlen(str) + 2) * sizeof(char));
-	while (i < ft_strlen(str))
+	while (i < (int)ft_strlen(str))
 	{
 		res[i] = str[i];
 		i++;
